+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTVWFZ2PD4CV5KXGMF9H641"
+        "01KPW92EH2BC70GXENMX17QVEB"
     )
     r"""The unique identifier of the data source"""