+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTVWFZ58VZ46FP0NQAD07PR"
+        "01KPW92EH664KWHHW8WE4J3PGY"
     )
     r"""The unique identifier of the data source"""