+++ packages/orq-rc/RELEASES.md
@@ -7259,3 +7259,13 @@ Based on:
 - [python v4.8.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.49 - packages/orq-rc
+
+## 2026-04-23 04:23:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.50 - packages/orq-rc